% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crgt.R
\docType{package}
\name{crgt}
\alias{crgt}
\alias{crgt-package}
\title{"crgt: data processing and analysis for the cued Rat Gambling Task."}
\description{
"This package provides functions for processing and analysis of operant
behavioural data for the cued and uncued versions of the Rat Gambling Task
(RGT) and the cRGT training version of the five choice serial reaction time
task (5CSRT). Provided functions enable the user to parse raw MedPC data
files, pivot them between long and wide formats, enabling them to perform
common statistical analyses for these behavioural tasks as used in the
Winstanley laboratory at the University of British Columbia. This package
aims to facilitate reproducible and standardized analysis of RGT data. Note
This package is currently in development and its functionality will be
improved over time."
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huttoncp/crgt}
  \item Report bugs at \url{https://github.com/crgt/issues}
}

}
