% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_by_session.R
\name{fcsrt_pivot}
\alias{fcsrt_pivot}
\title{Convert a session-aggregated version of data from the 5 choice serial
reaction time task (5CSRT; Robbins, 2002) that has been parsed by
\code{\link[=fcsrt_read]{fcsrt_read()}} or \code{\link[=fcsrt_read_file]{fcsrt_read_file()}} and prepared by
\code{\link[=fcsrt_prep]{fcsrt_prep()}} from wide format to long format or vice versa based
on its current shape. For more information on pivoting data frames between
long and wide formats, For more information on pivoting data frames between long and
wide formats, see the
\href{https://tidyr.tidyverse.org/articles/pivot.html}{pivoting} article on the
tidyr website.}
\usage{
fcsrt_pivot(.df)
}
\arguments{
\item{.df}{A data frame of 5CSRT data that has been imported into R using
either \code{\link[=fcsrt_read]{fcsrt_read()}} or \code{\link[=fcsrt_read_file]{fcsrt_read_file()}}.}
}
\value{
A data frame containing analysis-ready 5CSRT data with the following
columns, for the long format/shape (one column per variable):

\itemize{
\item \strong{subject:} rat ID number
\item \strong{session_date:} session date (yyyy-mm-dd) according to the MedPC file
\item \strong{session:} session number (subject's perspective)
\item \strong{n_trials:} number of trials initiated in the session
\item \strong{premat_n:} number of premature responses in the session
\item \strong{premat_prop:} premat_n/n_trials = premature response rate (for the session)
\item \strong{omissions:} number of omissions in the session (where the rat did not chose an option)
\item \strong{total_choices:} number of choices made in the session (successfully completed trials)
\item \strong{n_incorrect:} number of incorrect choices
\item \strong{n_correct:} number of correct choices
\item \strong{prop_correct:} n_correct/total_choices
\item \strong{stage:} The stage of the 5CSRT the rat did for the session
\item \strong{crierion_met:} TRUE if the rat achieved the 5CSRT stage's criterion performance level; FALSE if not
\item \strong{final_criterion_met:} TRUE if the rat achieved the final 5CSRT stage criterion performance level; FALSE if not
\item \strong{sessions_to_crietrion:} number of sessions the rat took to pass the final 5CSRT stage (if it passed)
}

The wide format/shape contains the same session-aggregated variables that
have been pivoted by session so that all repeated measurements by subject
appear in additional columns with names ending with with "*_s#" where "#" =
the session number, e.g., session_date_s1 is the date of that subject's first
session. The wide form of the data has one row per subject.
}
\description{
Convert a session-aggregated version of data from the 5 choice serial
reaction time task (5CSRT; Robbins, 2002) that has been parsed by
\code{\link[=fcsrt_read]{fcsrt_read()}} or \code{\link[=fcsrt_read_file]{fcsrt_read_file()}} and prepared by
\code{\link[=fcsrt_prep]{fcsrt_prep()}} from wide format to long format or vice versa based
on its current shape. For more information on pivoting data frames between
long and wide formats, For more information on pivoting data frames between long and
wide formats, see the
\href{https://tidyr.tidyverse.org/articles/pivot.html}{pivoting} article on the
tidyr website.
}
\references{
Robbins, T. (2002). The 5-choice serial reaction time task: behavioural
pharmacology and functional neurochemistry. Psychopharmacology, 163(3),
362-380.
}
\seealso{
\link[tidyr]{pivot_wider}, \link[tidyr]{pivot_longer}
}
