% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_by_session.R
\name{fcsrt_prep}
\alias{fcsrt_prep}
\title{Prepare a session-aggregated version of 5CSRT data for analysis.}
\usage{
fcsrt_prep(
  .df,
  shape = c("long", "wide"),
  output_file = NULL,
  recode_session_by_date = TRUE
)
}
\arguments{
\item{.df}{A data frame of 5CSRT data that has been imported into R using
either \code{\link[=fcsrt_read]{fcsrt_read()}} or \code{\link[=fcsrt_read_file]{fcsrt_read_file()}}.}

\item{shape}{Aggregation format or shape to use. The default, "long" format,
organizes the data with as one column per variable and represents repeated
measurements for experimental subjects as rows. This "tidy" format is
commonly required for analyses performed using R. The alternative "wide"
format instead uses additional columns to represent repeated measurements
for experimental subjects. The wide format is commonly required for
analyses performed using other programs like SPSS.}

\item{output_file}{Comma-separated variable (csv) file path and name to
export the combined data to, which should end in ".csv". e.g.
"5CSRT_data.csv" to save the file in your working directory.}

\item{recode_session_by_date}{If TRUE (default) re-codes the session variable
based on the session dates for each rat, in chronological order (from the
perspective of the rat). Set this to FALSE if you want to retain the
original session numbers from the raw MedPC file.}
}
\value{
A data frame containing analysis-ready 5CSRT data with the following columns

\itemize{
\item \strong{subject:} rat ID number
\item \strong{session_date:} session date (yyyy-mm-dd) according to the MedPC file
\item \strong{session:} session number (subject's perspective)
\item \strong{n_trials:} number of trials initiated in the session
\item \strong{premat_n:} number of premature responses in the session
\item \strong{premat_prop:} premat_n/n_trials = premature response rate (for the session)
\item \strong{omissions:} number of omissions in the session (where the rat did not chose an option)
\item \strong{total_choices:} number of choices made in the session (successfully completed trials)
\item \strong{n_incorrect:} number of incorrect choices
\item \strong{n_correct:} number of correct choices
\item \strong{prop_correct:} n_correct/total_choices
\item \strong{stage:} The stage of the 5CSRT the rat did for the session
\item \strong{crierion_met:} TRUE if the rat achieved the 5CSRT stage's criterion performance level; FALSE if not
\item \strong{final_criterion_met:} TRUE if the rat achieved the final 5CSRT stage criterion performance level; FALSE if not
\item \strong{sessions_to_crietrion:} number of sessions the rat took to pass the final 5CSRT stage (if it passed)
}
}
\description{
Prepare an analytical version of 5 choice serial reaction time task (5CSRT;
Robbins, 2002) data that has been parsed by \code{\link[=fcsrt_read]{fcsrt_read()}} or
\code{\link[=fcsrt_read_file]{fcsrt_read_file()}}. Either the long or wide forms of the data can be
generated and there is a convenience option to export the data to a csv file.
}
\references{
Robbins, T. (2002). The 5-choice serial reaction time task: behavioural
pharmacology and functional neurochemistry. Psychopharmacology, 163(3),
362-380.
}
