% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_medpc_files.R
\name{fcsrt_read_file}
\alias{fcsrt_read_file}
\title{Read a raw MedPC file from the five choice serial reaction time task.}
\usage{
fcsrt_read_file(path = "")
}
\arguments{
\item{path}{Character string for the raw MedPC file path, e.g.
"!2021-11-13_7h01m.Subject 1" if the file is in your working directory. A
file explorer window will pop-up to allow you to choose the file
interactively if the path is unspecified.}
}
\value{
A data frame containing 5CSRT data with the following columns

\itemize{
\item \strong{msn:} task version
\item \strong{start_date:} the session start date
\item \strong{start_time:} the session start time
\item \strong{subject:} subject identifier
\item \strong{box:} the operant box used to collect the data
\item \strong{comment:} any comments that were noted in MedPC
\item \strong{session:} session number
\item \strong{stim_dur:} stimulus duration
\item \strong{iti_dur:} inter-trial-interval duration
\item \strong{trial:} trial number
\item \strong{offered:} the choice offered in the trial
\item \strong{chosen:} the option chosen in the trial
\item \strong{correct_latency:} response time when correct
\item \strong{incorrect_latency:} response time when incorrect
\item \strong{collection_latency:} time to collect a reward
\item \strong{omission:} indicator of failure to respond
\item \strong{prematures:} premature response indicator
\item \strong{persev_r1:} perseveration
\item \strong{persev_r2:} perseveration
\item \strong{persev_r3:} perseveration
\item \strong{persev_r4:} perseveration
\item \strong{persev_r5:} perseveration
\item \strong{traypokes:} tray poke count
}
}
\description{
Read a raw MedPC (Tatham & Zurn, 1989) data file from the 5 choice serial
reaction time task (5CSRT; Robbins, 2002) and convert it into an R data.table.
}
\references{
Robbins, T. (2002). The 5-choice serial reaction time task: behavioural
pharmacology and functional neurochemistry. Psychopharmacology, 163(3),
362-380.

Tatham, T. A., & Zurn, K. R. (1989). The MED-PC experimental apparatus
programming system. Behavior Research Methods, Instruments, & Computers,
21(2), 294-302.
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
