% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_medpc_files.R
\name{iti5_read_file}
\alias{iti5_read_file}
\title{Read a raw MedPC file from the cued or uncued rat gambling task.}
\usage{
iti5_read_file(path = "")
}
\arguments{
\item{path}{Character string for the raw MedPC file path, e.g.
"!2021-11-13_7h01m.Subject 1" if the file is in your working directory. A
file explorer window will pop-up to allow you to choose the file
interactively if the path is unspecified.}
}
\value{
A data frame containing 5CSRT data with the following columns
\itemize{
\item \strong{msn:} task reward contingency coding version (A or B)
\item \strong{start_date:} the session start date
\item \strong{start_time:} the session start time
\item \strong{subject:} subject identifier
\item \strong{group:} experimental group
\item \strong{box:} the operant box used to collect the data
\item \strong{experiment:} experiment identifier
\item \strong{session:} session number
\item \strong{trial:} trial number
\item \strong{rewarded:} 1 if rewarded, 0 if punished
\item \strong{pellets:} number of pellets dispensed
\item \strong{omission:} indicator of failure to respond
\item \strong{chosen:} the option chosen in the trial
\item \strong{choice_lat:} time taken to make a choice
\item \strong{collect_lat:} time to collect a reward
\item \strong{pun_persev_h1:} perseverant response on hole 1 during punishment
\item \strong{pun_persev_h2:} perseverant response on hole 2 during punishment
\item \strong{pun_persev_h3:} perseverant response on hole 3 during punishment
\item \strong{pun_persev_h4:} perseverant response on hole 4 during punishment
\item \strong{pun_persev_h5:} perseverant response on hole 5 during punishment
\item \strong{pun_head_entry:} indicator of the rat responding during a time-out
\item \strong{pun_dur:} timeout punishment duration
\item \strong{premature_resp:} premature response indicator
\item \strong{premature_hole:} hole chosen when prematurely responding
\item \strong{rew_persev_h1:} perseverant response on hole 1 during reward
\item \strong{rew_persev_h2:} perseverant response on hole 1 during reward
\item \strong{rew_persev_h3:} perseverant response on hole 1 during reward
\item \strong{rew_persev_h4:} perseverant response on hole 1 during reward
\item \strong{rew_persev_h5:} perseverant response on hole 1 during reward
}
}
\description{
Read a raw MedPC (Tatham & Zurn, 1989) data file from the cued (Barrus &
Winstanley, 2016) or uncued (Zeeb et al., 2009) versions of the rat gambling
task (RGT) and convert it into an R data.table.
}
\references{
Barrus, M.M., & Winstanley, C.A. (2016). Dopamine D3 receptors modulate the
ability of win-paired cues to increase risky choice in a rat gambling task.
Journal of Neuroscience, 36(3), 785-794.

Tatham, T.A., & Zurn, K.R. (1989). The MED-PC experimental apparatus
programming system. Behavior Research Methods, Instruments, & Computers,
21(2), 294-302.

Zeeb, F.D., Robbins, T.W., & Winstanley, C.A. (2009). Serotonergic and
dopaminergic modulation of gambling behavior as assessed using a novel rat
gambling task. Neuropsychopharmacology, 34(10), 2329-2343.
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
