% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.contr.sum.R
\name{named.contr.sum}
\alias{named.contr.sum}
\title{Named Sum-to-Zero Contrasts}
\usage{
named.contr.sum(x, ...)
}
\arguments{
\item{x}{The input vector or factor.}

\item{...}{Additional arguments to be passed to the \code{contr.sum} function.}
}
\value{
A sum-to-zero contrast matrix.
}
\description{
Generates and labels sum-to-zero contrasts for factors.
}
\details{
If \code{x} is a factor, the function uses the levels of the factor
to compute sum-to-zero contrasts to support use of Type III Sums of Squares
for Analysis of Variance.

If \code{x} is a numeric vector with length 1, an error is thrown, indicating that it is not possible to
create names with an integer value and factor levels should be passed instead.
}
\examples{
x <- factor(c("A", "B", "A", "C"))
named.contr.sum(x)

}
\references{
Code was obtained from \href{https://stackoverflow.com/a/24516650}{https://stackoverflow.com/a/24516650}
}
\author{
Matthew Flickinger, \href{http://www.matthewflickinger.com/}{http://www.matthewflickinger.com/}
}
