% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_by_session.R
\name{rgt_pivot}
\alias{rgt_pivot}
\title{Convert a session-aggregated version of data from the cued (Barrus &
Winstanley, 2016) or uncued (Zeeb et al., 2009) rat gambling task (RGT) that
has been parsed by \code{\link[=rgt_read]{rgt_read()}} or \code{\link[=rgt_read_file]{rgt_read_file()}} and prepared by
\code{\link[=rgt_prep]{rgt_prep()}} from wide format to long format or vice versa based on its
current shape. For more information on pivoting data frames between long and
wide formats, see the
\href{https://tidyr.tidyverse.org/articles/pivot.html}{pivoting} article on the
tidyr website.}
\usage{
rgt_pivot(.df)
}
\arguments{
\item{.df}{A data frame of RGT data that has been imported into R using
either \code{\link[=rgt_read]{rgt_read()}} or \code{\link[=rgt_read_file]{rgt_read_file()}} and
prepared by \code{\link[=rgt_prep]{rgt_prep()}}.}
}
\value{
A data frame containing session-aggregated RGT data with the
following columns, for the long format/shape (one column per variable):

A data frame containing analysis-ready rGT data with columns that
depend on the source data MSNs. The long format/shape (one column per
variable) includes the following columns.

For standard cued/uncued ITI5 data:

\itemize{
\item \strong{subject:} rat ID number
\item \strong{group:} experimental group (only valid if entered into MedPC)
\item \strong{session_date:} session date (yyyy-mm-dd) according to the MedPC file
\item \strong{session:} session number (subject's perspective)
\item \strong{msn:} the MedPC program that was run during the session
\item \strong{n_trials:} number of trials initiated in the session
\item \strong{premat_n:} number of premature responses in the session
\item \strong{premat_prop:} premat_tot/n_trials = premature response rate (for the session)
\item \strong{total_choices:} number of choices made in the session (successfully completed trials)
\item \strong{P1:} proportion of total_choices where option P1 was selected (P1 preference)
\item \strong{P2:} proportion of total_choices where option P2 was selected (P2 preference)
\item \strong{P3:} proportion of total_chocies where option P3 was selected (P3 preference)
\item \strong{P4:} proportion of total_chocies where option P4 was selected (P4 preference)
\item \strong{choice_score:} overall choice preference score = (P1 + P2) - (P3 + P4)
\item \strong{omissions:} number of omissions in the session (where the rat did not chose an option)
\item \strong{mean_choice_lat:} mean time to choose an option on trials in the session when any choice (P1, P2, P3, or P4) was made
\item \strong{choice_lat_P1:} mean choice latency for trials when P1 was chosen
\item \strong{choice_lat_P2:} mean choice latency for trials when P2 was chosen
\item \strong{choice_lat_P3:} mean choice latency for trials when P3 was chosen
\item \strong{choice_lat_P4:} mean choice latency for trials when P4 was chosen
\item \strong{mean_collect_lat:} mean time to nose poke for sugar pellets when the rat was rewarded
}

ITI-9 data also includes:

\itemize{
\item \strong{mean_premat_time:} mean time of premature responses during the session
}

DD-rGT data also includes:

\itemize{
\item \strong{mean_premat_time:} mean time of premature responses during the session
\item \strong{dd_phenotype:} the targeted delay discounting phenotype (ideal or worsening)
\item \strong{iti_dur_P1:} ITI duration used for P1 during the session
\item \strong{iti_dur_P2:} ITI duration used for P2 during the session
\item \strong{iti_dur_P3:} ITI duration used for P3 during the session
\item \strong{iti_dur_P4:} ITI duration used for P4 during the session
\item \strong{mean_choice_iti:} mean ITI of options chosen during the session
}

The wide format/shape contains the same session-aggregated variables that
have been pivoted by session so that all repeated measurements by subject
appear in additional columns with names ending with with "*_s#" where "#" =
the session number, e.g., session_date_s1 is the date of that subject's first
session. The wide form of the data has one row per subject.
}
\description{
Convert a session-aggregated version of data from the cued (Barrus &
Winstanley, 2016) or uncued (Zeeb et al., 2009) rat gambling task (RGT) that
has been parsed by \code{\link[=rgt_read]{rgt_read()}} or \code{\link[=rgt_read_file]{rgt_read_file()}} and prepared by
\code{\link[=rgt_prep]{rgt_prep()}} from wide format to long format or vice versa based on its
current shape. For more information on pivoting data frames between long and
wide formats, see the
\href{https://tidyr.tidyverse.org/articles/pivot.html}{pivoting} article on the
tidyr website.
}
\references{
Barrus, M.M., & Winstanley, C.A. (2016). Dopamine D3 receptors modulate the
ability of win-paired cues to increase risky choice in a rat gambling task.
Journal of Neuroscience, 36(3), 785-794.

Zeeb, F.D., Robbins, T.W., & Winstanley, C.A. (2009). Serotonergic and
dopaminergic modulation of gambling behavior as assessed using a novel rat
gambling task. Neuropsychopharmacology, 34(10), 2329-2343.
}
\seealso{
\link[tidyr]{pivot_wider}, \link[tidyr]{pivot_longer}
}
