% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_by_session.R
\name{rgt_prep}
\alias{rgt_prep}
\title{Prepare a session-aggregated version of cued or uncued RGT data for analysis.}
\usage{
rgt_prep(
  .df,
  shape = c("long", "wide"),
  output_file = NULL,
  recode_session_by_date = TRUE
)
}
\arguments{
\item{.df}{A data frame of 5CSRT data that has been imported into R using
either \code{\link[=rgt_read]{rgt_read()}} or \code{\link[=rgt_read_file]{rgt_read_file()}}.}

\item{shape}{Aggregation format or shape to use. The default, "long" format,
organizes the data with as one column per variable and represents repeated
measurements for experimental subjects as rows. This "tidy" format is
commonly required for analyses performed using R. The alternative "wide"
format instead uses additional columns to represent repeated measurements
for experimental subjects (with one row per subject). The wide format is
commonly required for analyses performed using other programs like SPSS.}

\item{output_file}{Comma-separated variable (csv) file path and name to
export the combined data to, which should end in ".csv". e.g.
"rgt_data.csv" to save the file in your working directory.}

\item{recode_session_by_date}{If TRUE (default) re-codes the session variable
based on the session dates for each rat in chronological order (from the
perspective of the rat). Set this to FALSE if you want to retain the
original session numbers from the raw MedPC file.}
}
\value{
A data frame containing analysis-ready RGT data with the following columns

\itemize{
\item \strong{subject:} rat ID number
\item \strong{group:} experimental group (only valid if entered into MedPC)
\item \strong{session_date:} session date (yyyy-mm-dd) according to the MedPC file
\item \strong{session:} session number (subject's perspective)
\item \strong{n_trials:} number of trials initiated in the session
\item \strong{premat_n:} number of premature responses in the session
\item \strong{premat_prop:} premat_tot/n_trials = premature response rate (for the session)
\item \strong{total_choices:} number of choices made in the session (successfully completed trials)
\item \strong{P1:} proportion of total_choices where option P1 was selected (P1 preference)
\item \strong{P2:} proportion of total_choices where option P2 was selected (P2 preference)
\item \strong{P3:} proportion of total_chocies where option P3 was selected (P3 preference)
\item \strong{P4:} proportion of total_chocies where option P4 was selected (P4 preference)
\item \strong{choice_score:} overall choice preference score = (P1 + P2) - (P3 + P4)
\item \strong{omissions:} number of omissions in the session (where the rat did not chose an option)
\item \strong{mean_choice_lat:} mean time to choose an option on trials in the session when any choice (P1, P2, P3, or P4) was made
\item \strong{choice_lat_P1:} mean choice latency for trials when P1 was chosen
\item \strong{choice_lat_P2:} mean choice latency for trials when P2 was chosen
\item \strong{choice_lat_P3:} mean choice latency for trials when P3 was chosen
\item \strong{choice_lat_P4:} mean choice latency for trials when P4 was chosen
\item \strong{mean_collect_lat:} mean time to nose poke for sugar pellets when the rat was rewarded
}
}
\description{
Prepare an analytical version of session-aggregated data from the cued
(Barrus & Winstanley, 2016) or uncued (Zeeb et al., 2009) version of the rat
gambling task (RGT) that has been parsed by \code{\link[=rgt_read]{rgt_read()}} or
\code{\link[=rgt_read_file]{rgt_read_file()}}. Either the long or wide forms of the data can be generated
and there is a convenience option to export the data to a csv file.
}
\references{
Barrus, M.M., & Winstanley, C.A. (2016). Dopamine D3 receptors modulate the
ability of win-paired cues to increase risky choice in a rat gambling task.
Journal of Neuroscience, 36(3), 785-794.

Zeeb, F.D., Robbins, T.W., & Winstanley, C.A. (2009). Serotonergic and
dopaminergic modulation of gambling behavior as assessed using a novel rat
gambling task. Neuropsychopharmacology, 34(10), 2329-2343.
}
