% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_medpc_files.R
\name{rgt_read}
\alias{rgt_read}
\title{Read and combine a folder of raw MedPC files from the cued or uncued rat gambling task.}
\usage{
rgt_read(
  path = "",
  output_file = NULL,
  pattern = NULL,
  fixed = FALSE,
  negate = FALSE,
  progress = TRUE
)
}
\arguments{
\item{path}{Character string for path to the folder containing the raw
MedPC files, e.g. "~/raw_rgt_files/". A file explorer window will pop-up
to allow you to choose the folder interactively if the path is unspecified.}

\item{output_file}{Comma-separated variable (csv) file path and name to
export the combined data to, which should end in ".csv". e.g.
"rgt_data.csv" to save the file in your working directory.}

\item{pattern}{If the specified folder contains non-MedPC files or other
folders, this argument allows you to use a regular expression or fixed
string to tell the function which files to read. For example, raw MedPC
files often start with "!", so you might use pattern = "!" to only read
files containing a "!" in their name. See this \href{https://craig.rbind.io/post/2020-06-28-asgr-2-3-string-manipulation/}{this blog post}
to learn more about regular expressions in R.}

\item{fixed}{Set this to \code{TRUE} if you want the pattern argument to be
interpreted as a fixed string pattern (i.e. exactly as written) instead of
a regular expression for selecting the MedPC files in the chosen folder.
Ignored if "pattern" argument is not used.}

\item{negate}{Set this to \code{TRUE} to read files with names which do \emph{not}
match the specified pattern. Ignored if "pattern" argument is not used.}

\item{progress}{Should a progress bar be displayed when reading multiple
MedPC data files (default = TRUE)? See \link[purrr:progress_bars]{purrr::progress_bars}
for details and options.}
}
\value{
A data frame containing 5CSRT data with the following columns
\itemize{
\item \strong{msn:} task reward contingency coding version (A or B)
\item \strong{start_date:} the session start date
\item \strong{start_time:} the session start time
\item \strong{subject:} subject identifier
\item \strong{group:} experimental group
\item \strong{box:} the operant box used to collect the data
\item \strong{experiment:} experiment identifier
\item \strong{session:} session number
\item \strong{trial:} trial number
\item \strong{rewarded:} 1 if rewarded, 0 if punished
\item \strong{pellets:} number of pellets dispensed
\item \strong{omission:} indicator of failure to respond
\item \strong{chosen:} the option chosen in the trial
\item \strong{choice_lat:} time taken to make a choice
\item \strong{collect_lat:} time to collect a reward
\item \strong{pun_persev_h1:} perseverative response on hole 1 if punished
\item \strong{pun_persev_h2:} perseverative response on hole 2 if punished
\item \strong{pun_persev_h3:} perseverative response on hole 3 if punished
\item \strong{pun_persev_h4:} perseverative response on hole 4 if punished
\item \strong{pun_persev_h5:} perseverative response on hole 5 if punished
\item \strong{pun_head_entry:} indicator of the rat responding during a time-out
\item \strong{premature_resp:} premature response indicator
\item \strong{premature_hole:} hole chosen when prematurely responding
\item \strong{rew_persev_h1:} perseverative response on hole 1 if rewarded
\item \strong{rew_persev_h2:} perseverative response on hole 1 if rewarded
\item \strong{rew_persev_h3:} perseverative response on hole 1 if rewarded
\item \strong{rew_persev_h4:} perseverative response on hole 1 if rewarded
\item \strong{rew_persev_h5:} perseverative response on hole 1 if rewarded
}
}
\description{
Read a folder of raw MedPC (Tatham & Zurn, 1989) data files from the cued
(Barrus & Winstanley, 2016) or uncued (Zeeb et al., 2009) version of the rat
gambling task (RGT) and convert them into a single tibble with an option to
export the data to a comma-separated variable (.csv) file.
}
\references{
Barrus, M.M., & Winstanley, C.A. (2016). Dopamine D3 receptors modulate the
ability of win-paired cues to increase risky choice in a rat gambling task.
Journal of Neuroscience, 36(3), 785-794.

Tatham, T.A., & Zurn, K.R. (1989). The MED-PC experimental apparatus
programming system. Behavior Research Methods, Instruments, & Computers,
21(2), 294-302.

Zeeb, F.D., Robbins, T.W., & Winstanley, C.A. (2009). Serotonergic and
dopaminergic modulation of gambling behavior as assessed using a novel rat
gambling task. Neuropsychopharmacology, 34(10), 2329-2343.
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
