% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgt_stability.R
\name{rgt_rm_aov}
\alias{rgt_rm_aov}
\title{Check if rats' choice preferences are stable across recent sessions of the Rat Gambling Task (RGT)}
\usage{
rgt_rm_aov(.rgt_df, n_sessions = 5)
}
\arguments{
\item{.rgt_df}{A data frame of session-aggregated data from the cued or
uncued Rat Gambling Task. That has been parsed by \code{\link[=rgt_read]{rgt_read()}} or
\code{\link[=rgt_read_file]{rgt_read_file()}} and aggregated by session using \code{\link[=rgt_prep]{rgt_prep()}}.}

\item{n_sessions}{number of sessions to use to evaluate stability (default = 5).}
}
\value{
A repeated measures ANOVA results table
}
\description{
Use a repeated measures ANOVA to check if rats' choice preferences are stable
across the last "n" (5 by default) sessions of the cued (Barrus & Winstanley,
2016) or uncued (Zeeb et al., 2009) version of the rat gambling task (RGT)
that has been parsed by \code{\link[=rgt_read]{rgt_read()}} or \code{\link[=rgt_read_file]{rgt_read_file()}} and aggregated by
session using \code{\link[=rgt_prep]{rgt_prep()}}, in the long format (default format). In this
approach, the arcsine transformation is first applied to the choice
preference data because it is bounded by 0 and 1 and are actually binomial in nature (preference = # times chosen/total choices; MacDonald, 2009). N.B. this
approach is no longer considered best statistical practice, so this function
is mostly intended to provide a way to check for stability to enable
replication and comparisons to earlier publications. Current best practice is to use a logisitic mixed effects model to evaluate stability via \code{\link[=rgt_stability]{rgt_stability()}}
}
\details{
Choice preference stability of the sample can be evaluated by examining the
statistical significance of the session by choice interaction term.
}
\references{
Barrus, M.M., & Winstanley, C.A. (2016). Dopamine D3 receptors modulate the
ability of win-paired cues to increase risky choice in a rat gambling task.
Journal of Neuroscience, 36(3), 785-794.

Zeeb, F.D., Robbins, T.W., & Winstanley, C.A. (2009). Serotonergic and
dopaminergic modulation of gambling behavior as assessed using a novel rat
gambling task. Neuropsychopharmacology, 34(10), 2329-2343.

McDonald, J. H. (2009). Handbook of biological statistics (Vol. 2, pp. 6-59).
Baltimore, MD: sparky house publishing. URL: https://www.biostathandbook.com/transformation.html
}
