% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_crgt.R
\name{theme_crgt}
\alias{theme_crgt}
\title{theme_crgt}
\usage{
theme_crgt(
  base_size = 16,
  base_family = "sans",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  ...
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{...}{additional arguments to be passed to \code{\link[ggplot2]{theme}}}
}
\description{
This is a custom \href{https://ggplot2.tidyverse.org/}{ggplot2} theme for the
crgt package which controls all non-data display elements. Use
\code{\link[ggplot2]{theme}} instead if you just need to tweak the display of
an existing theme.
}
\references{
Wickham, H. (2016). ggplot2: elegant graphics for data analysis. New York, N.Y.: Springer-Verlag.
}
\seealso{
\code{\link[ggplot2]{theme}}
}
