% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgt_stability.R
\name{rgt_stability}
\alias{rgt_stability}
\title{Check if rats' choice preferences are stable across recent sessions of the Rat Gambling Task (RGT)}
\usage{
rgt_stability(
  .rgt_df,
  n_sessions = 5,
  alpha = 0.05,
  diagnostics = TRUE,
  n_sim = 1000
)
}
\arguments{
\item{.rgt_df}{A data frame of session-aggregated data from the cued or
uncued Rat Gambling Task. That has been parsed by \code{\link[=rgt_read]{rgt_read()}} or
\code{\link[=rgt_read_file]{rgt_read_file()}} and aggregated by session using \code{\link[=rgt_prep]{rgt_prep()}}.}

\item{n_sessions}{you may not have enough data to evaluate stability at the
individual rat level (random effects) if using fewer than 5 sessions of
data.}

\item{alpha}{Statistical significance level for p-values. Default is 0.05.}

\item{diagnostics}{Should model diagnostic checks be performed using
simulated residuals via \code{\link[DHARMa:simulateResiduals]{DHARMa::simulateResiduals()}} (TRUE/FALSE)? Default
is TRUE.}

\item{n_sim}{Number of simulations to use for diagnostic evaluation of model
fit via \code{\link[DHARMa:simulateResiduals]{DHARMa::simulateResiduals()}}. Default is 1,000.}
}
\value{
A list containing the following objects:

\itemize{
\item \strong{data:} A tibble (enhanced data frame) of the data used to fit the stability evaluation model
\item \strong{model:} The logistic GLMM used to evaluate choice preference stability fit with \code{\link[glmmTMB:glmmTMB]{glmmTMB::glmmTMB()}}
\item \strong{sim_resid:} Simulated model residuals using \code{\link[DHARMa:simulateResiduals]{DHARMa::simulateResiduals()}}. Can be graphed with plot().
\item \strong{anova:} Analysis of Variance for the model fixed effects using type III SS and sum-to-zero factor contrasts
\item \strong{random_slopes:} random slope coefficients and distribution metrics
\item \strong{random_slopes_plot:} ggplot2 graph of the random slopes for the effect of session for each subject facetted by choice option
\item \strong{posthoc_interaction:} Estimated marginal means and post-hoc pairwise contrasts statistics for the difference in linear session trends between choice options (only included if session x choice fixed effect interaction is significant)
\item \strong{posthoc_interaction_plot:} ggplot2 graph of the linear session trends for each choice option (only included if session x choice fixed effect interaction is significant)
}
}
\description{
Use a logistic mixed effects model to check if rats' choice preferences are
stable across the last "n" (5 by default) sessions of the cued (Barrus &
Winstanley, 2016) or uncued (Zeeb et al., 2009) version of the rat gambling
task (RGT) that has been parsed by \code{\link[=rgt_read]{rgt_read()}} or \code{\link[=rgt_read_file]{rgt_read_file()}} and
aggregated by session using \code{\link[=rgt_prep]{rgt_prep()}}, in the long format (default
format).
}
\details{
The \href{https://bbolker.github.io/mixedmodels-misc/glmmFAQ.html#model-specification}{formula} for the model is:

choice_prop ~ session + choice + session:choice + (session | subject/choice)

Where the dependent variable on the left-hand side, "choice_prop", is the
proportion of trials in which the rat chose a particular option ("choice"
variable), considered to represent its "preference" for that option. Because
this is a proportion out of a known total number of trials, we use a logistic
model for it by setting family = "binomial" for \code{\link[glmmTMB:glmmTMB]{glmmTMB::glmmTMB()}}, which
is the function we use to fit the model. \code{\link[glmmTMB:glmmTMB]{glmmTMB::glmmTMB()}} is preferred
over \code{\link[lme4:glmer]{lme4::glmer()}} because it tends to converge more often with fewer
warning/error messages in my experience.

Choice preference stability can be evaluated at a whole-sample level by
examining the statistical significance of the fixed effect terms "session"
and "session:choice" (the session by choice interaction) on the right-hand
side of the model formula. The stability of RGT choice preferences for each
rat can be evaluated by examining the random slope parameters, or the
"(session | subject/choice)" part of the model formula, where we want to see
if there is a change in choice preference across sessions for each choice
option presented to each subject.
}
\references{
Barrus, M.M., & Winstanley, C.A. (2016). Dopamine D3 receptors modulate the
ability of win-paired cues to increase risky choice in a rat gambling task.
Journal of Neuroscience, 36(3), 785-794.

Zeeb, F.D., Robbins, T.W., & Winstanley, C.A. (2009). Serotonergic and
dopaminergic modulation of gambling behavior as assessed using a novel rat
gambling task. Neuropsychopharmacology, 34(10), 2329-2343.
}
